First pass at adding an option to select DSM input type
Because DSM packets do not have checksums it is possible for the DSM decoder to find 'valid' packets in XBus data, decode them, and override XBus outputs. To avoid this behaviour, an option has been added to choose (currently at compile time) which input decoding to use.

Ideally, this option should be runtime selectable